Fix bug where on the first load of the website the data was being accessed before it was fetched.
The function had 2 outcomes: the data was present in localstorage
and therefor treated as sync and another where the data had to be fetched and
therefor was async.

The solution was to treat all return values as async and chain the DOM
manipulation functions to the resolution of the data fetch promise.